Let's focus on the edit form first - it'll be a little bit easier to get working. Go to /admin/article and click to edit one of the existing articles. So, based on the location, we need to make this specificLocationName field have different options.
Determining the specificLocationName Choices

Open ArticleFormType and go to the bottom. I'm going to paste in a function I wrote called getLocationNameChoices(). You can copy this function from the code block on this page. But, it's fairly simple: We pass it the  string, which will be one of solar_system, star or interstellar_space, and it returns the choices for the specificLocationName field. If we choose 'solar system', it returns planets. If we choose 'star', it returns some popular stars. And if we choose 'Interstellar space', it returns null, because we actually don't want the drop-down to be displayed at all in that case.
Oh, and I'm using array_combine() just because I want the display values and the values set back on my entity to be the same. This is equivalent to saying 'Mercury' => 'Mercury'... but saves me some duplication.