We can use the submitted data to dynamically change the specificLocationName field to use the correct choices, right before validation occurs. Actually, this hook happens after validation - but we'll use a trick where we remove and re-add the field, to get around this.
To start, create a new private function called setupSpecificLocationNameField(). The job of this function will be to dynamically add the specificLocationName field with the correct choices. It will accept a FormInterface - we'll talk about that in a minute - and a ?string $ location, the ? part so this can be null.

Inside, first check if  is null. If it is, take the  object and actually ->remove() the specificLocationName field and return. Here's the idea: if when I originally rendered the form there was a location set, then, thanks to our logic in buildForm(), there will be a specificLocationName field. But if we changed it to 'Choose a location', meaning we are not selecting a location, then we want to remove the specificLocationName field before we do any validation. We're kind of trying to do the same thing in here that our future JavaScript will do instantly on the frontend: when we change to 'Choose a location' - we will want the field to disappear.

Next, get the  by using $ this->getLocationNameChoices() and pass that $ location. Then, similar to above, if (null === $ choices) remove the field and return. This is needed for when the user selects 'Interstellar Space': that doesn't have any specific location name choices, and so we don't want that field at all.

Finally, we do want the specificLocationName field, but we want to use our new choices. Scroll up and copy the $ builder->add() section for this field, paste down here, and change  to $ form - these two objects have an identical add() method. For choices pass $ choices.