We created setupSpecificLocationNameField function so that we can call it from inside of our listener callback. Start with $ form = $ event->getForm(): that gives us the actual Form object for this one field. Now call $ this->setupSpecificLocationNameField() and, for the first argument, pass it $ form->getParent().
This is tricky. The $ form variable is the Form object that represents just the location field. But we want to pass the top level Form object into the function so that the specificLocationName field can be added or removed from it.

The second argument is the location itself, which will be $ form->getData(), or $ event->getData().

Okay guys, I know this is craziness, but we're ready to try it! Refresh to resubmit the form. It saves. Now change the Location to 'Near a Star'. In a few minutes, our JavaScript will reload the specificLocationName field with the new options. To fake that, inspect the element. Let's go copy a real star name - how about Sirius. Change the selected option's value to that string.

Hit update! Yes! It saved! We were able to change both the location and specificLocationName fields at the same time.