JS Setup: Adding data- Attributes & Classes
Next, open _form.html.twig. Our JavaScript will need to be able to find the location select element so it can read its value and the specificLocationName field so it can replace its contents. It also needs to know the URL to our new endpoint.

No problem: for the location field, pass an attr array variable. Add a data-specific-location-url key set to path('admin_article_location'). Then, add a class set to js-article-form-location.

Next, surround the specificLocationName field with a new <div class='js-specific-location-target'>. I'm adding this as a new element around the field instead of on the select element so that we can remove the field without losing this target element.