Adding the JavaScript
Ok, we're ready for the JavaScript! Open up the public/ directory and create a new file: admin_article_form.js. I'm going to paste in some JavaScript that I prepped: you can copy this from the code block on this page.

Before we talk about the specifics, let's include this with the script tag. Unfortunately, we can't include JavaScript directly in _form.html.twig because that's an included template. So, in the edit template, override {% block javascripts %}, call the {{ parent() }} function and then add a <script> tag with src='{{ asset('js/admin_article_form.js') }}. Copy that, open the new template, and paste this at the bottom of the javascripts block.

Before we try this, let's check out the JavaScript so we can see the entire flow. I made the code here as simple, and unimpressive as possible - but it gets the job done. First, we select the two elements: $ locationSelect is the actual select element and $ specificLocationTarget represents the div that's around that field. The $ on the variables is meaningless - I'm just using it to indicate that these are jQuery elements.

Next, when the location select changes, we make the AJAX call by reading the data-specific-location-url attribute. The location key in the data option will cause that to be set as a query parameter.

Finally, on success, if the response is empty, that means that we've selected an option that should not have a specificLocationName dropdown. So, we look inside the $ specificLocationTarget for the select and remove it to make sure it doesn't submit with the form. On the wrapper div, we also need to add a Bootstrap class called d-none: that stands for display none. That will hide the entire element, including the label.

If there is some HTML returned, we do the opposite: replace the entire HTML of the target with the new HTML and remove the class so it's not hidden. And... that's it!

There are a lot of moving pieces, so let's try it! Refresh the edit page. The current location is 'star' and... so far, no errors in my console. Change the option to 'The Solar System'. Yes! The options updated! Try 'Interstellar Space'... gone!

If you look deeper, the js-specific-location-target div is still there, but it's hidden, and only has the label inside. Change back to 'The Solar System'. Yep! The d-none is gone and it now has a select field inside.

Try saving: select 'Earth' and Update! We got it! We can keep changing this all day long - all the pieces are moving perfectly.

I'm super happy with this, but it is a complex setup - I totally admit that. If you have this situation, you need to choose the best solution: if you have a big form with 1 dependent field, what we just did is probably a good option. But if you have a small form, or it's even more complex, it might be better to skip the form component and code everything with JavaScript and API endpoints. The form component is a great tool - but not the best solution for every problem.

Next: there are a few small details we need to clean up before we are fully done with this form. Let's squash those!