When we change to the solar system, great! It loads the planets. We can even change to 'Interstellar Space' and it disappears. We're amazing! And when we change it to 'Choose a Location'... uhhh oh! Nothing happened? Ah, the Ajax part of the web debug toolbar is trying to tell me that there was a 500 error!
By the way, this is one of the coolest features of the web debug toolbar: when you get a 500 error on an AJAX call, you can click this link to jump straight into the profiler for that request! It takes us straight to the Exception screen so we can see exactly what we messed up, I mean, what went wrong... that may or may not be our fault.