Fixing our Empty Value Bug
Apparently ArticleFormType line 125 has an undefined 'empty string' index. Let's go check that out. This is the method that we call to get the correct specificLocationName choices. But, in this case, the location was an empty string, and that was a super not valid key.

Let's make sure that worked: on your browser, switch back to the solar system, and then back to 'Choose a Location'. Nice! The field disappears and no 500 error this time.

Forcing specificLocationName to null

There's one other subtle problem with our setup. To see it, refresh this page. In the database, this article's location is star, specificLocationName is Rigel and id is 28. Let's go verify this in the database: find your terminal and run:
php bin/console doctrine:query:sql 'SELECT * FROM article WHERE id = 28'
Yep! All the data looks like we expected! But now, change the location to 'Interstellar Space' and hit update. It works... but try that query again:
php bin/console doctrine:query:sql 'SELECT * FROM article WHERE id = 28'
Ok: the location is interstellar_space, but ah! The specific_location_name is still Rigel! This may or may not be a real problem - depending on how you use this data. But it's for sure technically wrong: when we change the location to interstellar_space, the specific_location_name should be set back to null: we are not at Rigel.

The reason this did not happen is subtle. When we change the location to 'Interstellar Space' and submit, our POST_SUBMIT listener function calls setupSpecificLocationNameField(), which sees that there are no choices for this location and so removes the field entirely. The end result is that the form makes no changes to the specificLocationName property on Article: it just never calls setSpecificLocationName() at all... because that field isn't part of the form!

That is the correct behavior. But, it means that we need to do a little bit more work to clean things up. There are a few ways to fix this inside the form itself. But, honestly, they're overly-complex.