First, add @IsGranted(ROLE_USER) to at least make sure the user is logged in. Then, inside the method, if not $ this->isGranted('ROLE_ADMIN_ARTICLE') and $ this->getUser()->getArticles() === 0, then we should not have access. Wait, but the ->getArticles() method is not auto-completing for me.
Oh, I know why! Go to the top of this class and change the base class from extends AbstractController to extends BaseController.

Reminder: BaseController is a controller that we created. It extends AbstractController but it adds a return type to getUser() with our User class so we get auto-completion.

Back down in our method, we can say $ this->getUser()->getArticles()->isEmpty(), which is a method on Doctrine's Collection object. So, if we don't have ROLE_ADMIN_ARTICLE and we are not the author of any articles, throw $ this->createAccessDeniedException().

Done! And just to make sure I didn't completely break things, if I change the location to 'Near a star'... yea! It still loads.

What really made adding this security easy was being able to call $ this->getUser()->getArticles(). The problem is that if this user is the author of 200 articles, then this will query for 200 rows of articles and hydrate those into 200 full objects, just to figure out that, yes, we are the author of at least one article. All we really need is a quick count query of the articles.